requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
alembic
pydantic
pydantic-settings
python-jose
passlib[bcrypt]
python-dotenv
authlib
requests
loguru
fastapi-pagination
apscheduler
pytest
httpx
pytest-cov
black
isort
ruff
email-validator
aiohttp
azure-storage-blob
azure-identity
python-multipart
PyJWT
psycopg2
psycopg2-binary